+++ graphics/vtkEncodedGradientEstimator.cxx
@@ -294,8 +294,8 @@ void vtkEncodedGradientEstimator::ComputeCircleLimits( int size )
     start = (start<0)?(0):(start);
     end   = (end>(size-1))?(size-1):(end);
 
-    *(ptr++) = start;
-    *(ptr++) = end;
+    *(ptr++) = (int) start;
+    *(ptr++) = (int) end;
     }
 }
