+++ imaging/vtkAxisActor2D.cxx
@@ -517,7 +517,7 @@ void vtkAxisActor2D::ComputeRange(float inRange[2], float outRange[2],
     lastInterval = interval;
     }
   interval = lastInterval;
-  numTicks = ceil(range/interval) + 1; //trim number of extra ticks
+  numTicks = (int) ceil(range/interval) + 1; //trim number of extra ticks
   
   // Get the start value of the range
   for ( j=0; j < VTK_NUM_DIVS; j++ )
@@ -547,7 +547,7 @@ int vtkAxisActor2D::SetFontSize(vtkViewport *viewport,
   int tempi[2], fontSize, target, length=(size[0]>size[1]?size[0]:size[1]);
 
   // find the best size for the font
-  target = 0.015*factor*size[0] + 0.015*factor*size[1];
+  target = (int)(0.015*factor*size[0] + 0.015*factor*size[1]);
 
   fontSize = target;
   textMapper->SetFontSize(fontSize);
@@ -593,8 +593,8 @@ void vtkAxisActor2D::SetOffsetPosition(float xTick[3], float theta,
   center[0] = xTick[0] + x*sin(theta);
   center[1] = xTick[1] - y*cos(theta);
     
-  pos[0] = center[0] - stringWidth/2.0;
-  pos[1] = center[1] - stringHeight/2.0;
+  pos[0] = (int)(center[0] - stringWidth/2.0);
+  pos[1] = (int)(center[1] - stringHeight/2.0);
   
   actor->SetPosition(pos[0], pos[1]);
 }
